+++ DependencyInjection/Configuration.php
@@ -153,19 +153,8 @@ final class Configuration implements ConfigurationInterface
                     ->children()
                         ->scalarNode('router')->defaultValue('router')->setDeprecated('The "%path%.%node%" configuration key has been deprecated in FOSRestBundle 2.8.')->end()
                         ->scalarNode('templating')
-                            ->defaultValue(static function () {
-                                @trigger_error('Not setting the "fos_rest.service.templating" configuration option to "null" is deprecated since FOSRestBundle 2.8.', E_USER_DEPRECATED);
-
-                                return 'templating';
-                            })
-                            ->validate()
-                                ->ifTrue(static function ($v) { return $v; })
-                                ->then(static function ($v) {
-                                    @trigger_error('Not setting the "fos_rest.service.templating" configuration option to "null" is deprecated since FOSRestBundle 2.8.', E_USER_DEPRECATED);
-
-                                    return $v;
-                                })
-                            ->end()
+                            ->defaultValue('templating')
+                            ->setDeprecated('The "%path%.%node%" option is deprecated since FOSRestBundle 2.8.')
                         ->end()
                         ->scalarNode('serializer')->defaultNull()->end()
                         ->scalarNode('view_handler')->defaultValue('fos_rest.view_handler.default')->end()