+++ Tests/DependencyInjection/ConfigurationTest.php
@@ -61,9 +61,6 @@ class ConfigurationTest extends TestCase
                     ],
                     'routing_loader' => false,
                     'body_listener' => false,
-                    'service' => [
-                        'templating' => null,
-                    ],
                     'view' => [
                         'default_engine' => null,
                     ],
@@ -94,9 +91,6 @@ class ConfigurationTest extends TestCase
             ],
             'routing_loader' => false,
             'body_listener' => false,
-            'service' => [
-                'templating' => null,
-            ],
             'view' => [
                 'default_engine' => null,
             ],
@@ -133,9 +127,6 @@ class ConfigurationTest extends TestCase
                     ],
                     'routing_loader' => false,
                     'body_listener' => false,
-                    'service' => [
-                        'templating' => null,
-                    ],
                     'view' => [
                         'default_engine' => null,
                     ],
@@ -161,9 +152,6 @@ class ConfigurationTest extends TestCase
                     ],
                     'routing_loader' => false,
                     'body_listener' => false,
-                    'service' => [
-                        'templating' => null,
-                    ],
                     'view' => [
                         'default_engine' => null,
                     ],
@@ -190,9 +178,6 @@ class ConfigurationTest extends TestCase
                     ],
                     'routing_loader' => false,
                     'body_listener' => false,
-                    'service' => [
-                        'templating' => null,
-                    ],
                     'view' => [
                         'default_engine' => null,
                     ],
@@ -225,9 +210,6 @@ class ConfigurationTest extends TestCase
                     ],
                     'routing_loader' => false,
                     'body_listener' => false,
-                    'service' => [
-                        'templating' => null,
-                    ],
                     'view' => [
                         'default_engine' => null,
                     ],