+++ Tests/DependencyInjection/FOSRestExtensionTest.php
@@ -87,9 +87,6 @@ class FOSRestExtensionTest extends TestCase
                     'serialize_exceptions' => false,
                 ],
                 'routing_loader' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -111,9 +108,6 @@ class FOSRestExtensionTest extends TestCase
                     'serialize_exceptions' => false,
                 ],
                'routing_loader' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -144,9 +138,6 @@ class FOSRestExtensionTest extends TestCase
                     'serialize_exceptions' => false,
                 ],
                 'routing_loader' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -176,9 +167,6 @@ class FOSRestExtensionTest extends TestCase
                     'serialize_exceptions' => false,
                 ],
                 'routing_loader' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -213,9 +201,6 @@ class FOSRestExtensionTest extends TestCase
                     'serialize_exceptions' => false,
                 ],
                 'routing_loader' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -245,9 +230,6 @@ class FOSRestExtensionTest extends TestCase
                 'format_listener' => false,
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -269,9 +251,6 @@ class FOSRestExtensionTest extends TestCase
                 ],
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -295,9 +274,6 @@ class FOSRestExtensionTest extends TestCase
                 ],
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -320,9 +296,6 @@ class FOSRestExtensionTest extends TestCase
                 'param_fetcher_listener' => true,
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -346,9 +319,6 @@ class FOSRestExtensionTest extends TestCase
                 'param_fetcher_listener' => 'force',
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -377,9 +347,6 @@ class FOSRestExtensionTest extends TestCase
                 ],
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -406,9 +373,6 @@ class FOSRestExtensionTest extends TestCase
                 ],
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -428,9 +392,6 @@ class FOSRestExtensionTest extends TestCase
                 ],
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -459,9 +420,6 @@ class FOSRestExtensionTest extends TestCase
                 ],
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -484,9 +442,6 @@ class FOSRestExtensionTest extends TestCase
                 ],
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -510,9 +465,6 @@ class FOSRestExtensionTest extends TestCase
                 ],
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -536,16 +488,13 @@ class FOSRestExtensionTest extends TestCase
                ],
                'routing_loader' => false,
                'body_listener' => false,
-               'service' => [
-                   'templating' => null,
-               ],
                'view' => [
                    'default_engine' => null,
                    'force_redirects' => [],
                ],
            ],
         ], $this->container);
-        $this->assertEquals(204, $this->container->getDefinition('fos_rest.view_handler.default')->getArgument(5));
+        $this->assertEquals(204, $this->container->getDefinition('fos_rest.view_handler.default')->getArgument(6));
     }
 
     public function testForceEmptyContentIs200()
@@ -558,9 +507,6 @@ class FOSRestExtensionTest extends TestCase
                 ],
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'empty_content' => 200,
@@ -569,7 +515,7 @@ class FOSRestExtensionTest extends TestCase
             ],
         ];
         $this->extension->load($config, $this->container);
-        $this->assertEquals(200, $this->container->getDefinition('fos_rest.view_handler.default')->getArgument(5));
+        $this->assertEquals(200, $this->container->getDefinition('fos_rest.view_handler.default')->getArgument(6));
     }
 
     public function testViewSerializeNullDefault()
@@ -582,16 +528,13 @@ class FOSRestExtensionTest extends TestCase
                 ],
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
                 ],
             ],
         ], $this->container);
-        $this->assertFalse($this->container->getDefinition('fos_rest.view_handler.default')->getArgument(6));
+        $this->assertFalse($this->container->getDefinition('fos_rest.view_handler.default')->getArgument(7));
     }
 
     public function testViewSerializeNullIsTrue()
@@ -604,9 +547,6 @@ class FOSRestExtensionTest extends TestCase
                 ],
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -615,7 +555,7 @@ class FOSRestExtensionTest extends TestCase
             ],
         ];
         $this->extension->load($config, $this->container);
-        $this->assertTrue($this->container->getDefinition('fos_rest.view_handler.default')->getArgument(6));
+        $this->assertTrue($this->container->getDefinition('fos_rest.view_handler.default')->getArgument(7));
     }
 
     public function testValidatorAliasWhenEnabled()
@@ -629,9 +569,6 @@ class FOSRestExtensionTest extends TestCase
                 ],
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -653,9 +590,6 @@ class FOSRestExtensionTest extends TestCase
                 ],
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -677,9 +611,6 @@ class FOSRestExtensionTest extends TestCase
                 'exception' => [
                     'exception_listener' => false,
                     'serialize_exceptions' => false,
-                ],
-                'service' => [
-                    'templating' => null,
                 ],
                 'view' => [
                     'default_engine' => null,
@@ -729,9 +660,6 @@ class FOSRestExtensionTest extends TestCase
                 ],
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -756,9 +684,6 @@ class FOSRestExtensionTest extends TestCase
                     'enabled' => false,
                 ],
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -785,9 +710,6 @@ class FOSRestExtensionTest extends TestCase
                     ],
                     'routing_loader' => [
                         'include_format' => false,
-                    ],
-                    'service' => [
-                        'templating' => null,
                     ],
                     'view' => [
                         'default_engine' => null,
@@ -830,9 +752,6 @@ class FOSRestExtensionTest extends TestCase
                     'routing_loader' => [
                         'default_format' => 'xml',
                     ],
-                    'service' => [
-                        'templating' => null,
-                    ],
                     'view' => [
                         'default_engine' => null,
                         'force_redirects' => [],
@@ -870,9 +789,6 @@ class FOSRestExtensionTest extends TestCase
                     'exception' => [
                         'exception_listener' => false,
                         'serialize_exceptions' => false,
-                    ],
-                    'service' => [
-                        'templating' => null,
                     ],
                     'view' => [
                         'default_engine' => null,
@@ -923,9 +839,6 @@ class FOSRestExtensionTest extends TestCase
                 ],
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -954,9 +867,6 @@ class FOSRestExtensionTest extends TestCase
                 ],
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -1000,9 +910,6 @@ class FOSRestExtensionTest extends TestCase
                 'exception' => $exceptionConfig,
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -1090,9 +997,6 @@ class FOSRestExtensionTest extends TestCase
                 ],
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -1114,9 +1018,6 @@ class FOSRestExtensionTest extends TestCase
                 ],
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -1142,9 +1043,6 @@ class FOSRestExtensionTest extends TestCase
                 ],
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -1167,9 +1065,6 @@ class FOSRestExtensionTest extends TestCase
                 ],
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -1230,9 +1125,6 @@ class FOSRestExtensionTest extends TestCase
                 ],
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -1262,9 +1154,6 @@ class FOSRestExtensionTest extends TestCase
                 ],
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -1285,9 +1174,6 @@ class FOSRestExtensionTest extends TestCase
                 ],
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -1307,9 +1193,6 @@ class FOSRestExtensionTest extends TestCase
             ],
             'routing_loader' => false,
             'body_listener' => false,
-            'service' => [
-                'templating' => null,
-            ],
             'view' => [
                 'default_engine' => null,
                 'force_redirects' => [],
@@ -1357,9 +1240,6 @@ class FOSRestExtensionTest extends TestCase
                 ],
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -1393,9 +1273,6 @@ class FOSRestExtensionTest extends TestCase
                 ],
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],
@@ -1423,9 +1300,6 @@ class FOSRestExtensionTest extends TestCase
                 ],
                 'routing_loader' => false,
                 'body_listener' => false,
-                'service' => [
-                    'templating' => null,
-                ],
                 'view' => [
                     'default_engine' => null,
                     'force_redirects' => [],